Working FFT in WASM

+++ src/audio/message.ts
@@ -23,14 +23,14 @@ export namespace Change {
 export namespace Time {
   export type Message = ReturnType<typeof make>;
   export const type = "time";
-  export const make = (buffer: Float32Array) => ({ type, buffer });
+  export const make = (buffer: Uint8Array) => ({ type, buffer });
   export const check = (message: any): message is Message => message?.type === type;
 }
 
 export namespace Frequency {
   export type Message = ReturnType<typeof make>;
   export const type = "frequency";
-  export const make = (buffer: Float32Array) => ({ type, buffer });
+  export const make = (buffer: Uint8Array) => ({ type, buffer });
   export const check = (message: any): message is Message => message?.type === type;
 }
